\name{testSAN-package}
\alias{testSAN-package}
\alias{testSAN}
\docType{package}
\title{\packageTitle{testSAN}}
\description{\packageDescription{testSAN}}
\section{Package Content}{\packageIndices{testSAN}}
\author{\packageAuthor{testSAN}}
\section{Maintainer}{\packageMaintainer{testSAN}}
\keyword{package}
